added midterm review readme
