+++ midtermReviewNotes.md
@@ -1,0 +1,12 @@
+### What was surprising or interesting about reviewing your partner’s codebase?
+- Implementing google maps with spring and javascript was similar to how i implemented the chart on my project.
+- Xia showed me a different way to search in the database by adding methods into the crud repo interface.
+
+### What was surprising or interesting about reviewing your own codebase?
+-Needed to java doc, document code.
+-In the header.html, make a div and add a th:if condition for when the user can view the nav as a logged in user or 
+not a logged in user.
+
+### What would you do differently next time?
+- Use the CRUD repository to query things instead of writing helper methods to gather expenses. 
+- Create a dashboard controller separately instead of putting it in the home controller.
